+++ fast_dash/__init__.py
@@ -2,7 +2,7 @@
 
 __author__ = """Kedar Dabhadkar"""
 __email__ = "kedar@fastdash.app"
-__version__ = "0.2.10"
+__version__ = "0.2.11"
 
 from fast_dash.Components import (
     Graph,
@@ -22,7 +22,8 @@ from fast_dash.Components import (
     Table
 )
 import dash
-from fast_dash.fast_dash import FastDash, fastdash
+from dash import Output, Input, State, callback, no_update
+from fast_dash.fast_dash import FastDash, fastdash, update, notify
 from fast_dash.utils import Fastify
 
 __all__ = [